Sem 4 NST & Rishihood/GenAI Lab/AI Agent Deployment/server/requirements.txt
fastapi==0.128.5
uvicorn[standard]==0.40.0

# LLM stack
langchain==1.2.9
langchain-core==1.2.9
langchain-groq==1.1.2
langgraph==1.0.8

# Env + HTTP
python-dotenv==1.2.1
requests==2.32.5

# Vector / embeddings (only if using RAG)
faiss-cpu==1.13.2
sentence-transformers==5.2.2

# Data handling (optional but useful)
pydantic==2.12.5

# Frontend (if using Streamlit)
streamlit==1.56.0